# api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Boxty API
  version: 1.0.0
  description: |
    Boxty — programmable sandbox platform. Run code in cloud with gVisor isolation,
    secrets, cron scheduling, image building, persistent volumes, and web endpoint API.
servers:
  - url: https://api.boxty.dev
    description: Production
  - url: http://localhost:3000
    description: Development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from login/register
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API key format "Bearer bxty_..."
  schemas:
...
    Sandbox:
      type: object
      properties:
        id: { type: string }
        status: { type: string, enum: [running, stopped, snapshotted] }
        url: { type: string, format: uri }
        image: { type: string }
        cpu: { type: integer }
        memory: { type: integer }
        gpu: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        timeout: { type: integer }
    ExecResult:
      type: object
      properties:
        stdout: { type: string }
        stderr: { type: string }
        exit_code: { type: integer }
        duration_ms: { type: integer }
...
    ImageBuild:
      type: object
      properties:
        id: { type: string }
        image_url: { type: string, nullable: true }
        status: { type: string, enum: [building, done, failed] }
        error: { type: string, nullable: true }
...
    Workspace:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        created_at: { type: string, format: date-time }
...
paths:
...
  /api/sandboxes:
    post:
      summary: Create sandbox
      tags: [Sandboxes]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image]
              properties:
                image: { type: string, example: "pytorch:latest" }
                cpu: { type: integer, default: 2 }
                memory: { type: integer, default: 4096 }
                gpu: { type: string, nullable: true }
                timeout: { type: integer, default: 3600 }
                disk_size_gb: { type: integer, default: 10 }
                volume: { type: string, nullable: true }
                volume_mount_path: { type: string, default: "/mnt/volume" }
      responses:
        '201':
          description: Sandbox created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Sandbox' }
    get:
      summary: List sandboxes
      tags: [Sandboxes]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      responses:
        '200':
          description: List of sandboxes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Sandbox' }
  /api/sandboxes/{id}:
    get:
      summary: Get sandbox details
      tags: [Sandboxes]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Sandbox details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Sandbox' }
    delete:
      summary: Stop and delete sandbox
      tags: [Sandboxes]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Sandbox stopped
  /api/sandboxes/{id}/exec:
    post:
      summary: Execute command in sandbox
      tags: [Sandboxes]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [command]
              properties:
                command: { type: string }
                timeout: { type: integer }
      responses:
        '200':
          description: Execution result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExecResult' }
  /api/sandboxes/{id}/snapshot:
    post:
      summary: Create snapshot
      tags: [Sandboxes]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
      responses:
        '200':
          description: Snapshot created
  /api/sandboxes/restore:
    post:
      summary: Restore from snapshot
      tags: [Sandboxes]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [snapshotKey, image]
              properties:
                snapshotKey: { type: string }
                image: { type: string }
                cpu: { type: integer }
                memory: { type: integer }
                gpu: { type: string }
                secrets: { type: array }
      responses:
        '201':
          description: Sandbox restored
...
  /api/workers:
    get:
      summary: List workers
      tags: [Workers]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Workers list
...
  /api/images/build:
    post:
      summary: Build image
      tags: [Images]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [base_image, commands]
              properties:
                base_image: { type: string }
                commands: { type: array, items: { type: string } }
                name: { type: string }
      responses:
        '201':
          description: Build started
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImageBuild' }
...
  /api/workspaces:
    post:
      summary: Create workspace
      tags: [Workspaces]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        '201':
          description: Workspace created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
    get:
      summary: List workspaces
      tags: [Workspaces]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      responses:
        '200':
          description: Workspaces list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Workspace' }
  /api/workspaces/{id}:
    get:
      summary: Get workspace
      tags: [Workspaces]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Workspace details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
    delete:
      summary: Delete workspace
      tags: [Workspaces]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Workspace deleted
...